OSNotificationOpenBehaviorFromPushPayload
* Split logic from GenerateNotificationOpenIntentFromPushPayload into
OSNotificationOpenBehaviorFromPushPayload so this can be used independly
in a follow up commit
